requirements.txt
streamlit 
openai 
python-docx
markdown2

